# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Documentation Quber",
        "description": "Documentation générale",
        "version": "0.1"
    },
    "servers": [{
        "url": "http://localhost:8090"
    }],
...
    "paths": {
...
        "/livreurs": {
            "post": {
                "tags": [
                    "Livreurs"
                ],
                "description": "Création d'un livreur",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/body"
                            },
                            "examples": {
                                "0": {
                                    "value": "{\r\n    \"nom\": \"Villeneuve\",\r\n    \"prenom\": \"Jacques\",\r\n    \"Voiture\": \"Williams\",\r\n    \"Quartier\": \"Westmont\"\r\n}"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retourne le json du livreur"
                    }
                },
                "servers": [{
                    "url": "http://localhost:8090"
                }]
            },
            "servers": [{
                "url": "http://localhost:8090"
            }]
        },
        "/usagers": {
            "post": {
                "tags": [
                    "Usager"
                ],
                "description": "Création d'un usager",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/body_1"
                            },
                            "examples": {
                                "0": {
                                    "value": "{\r\n    \"nom\": \"Frankenstein\",\r\n    \"prenom\": \"Dr\",\r\n    \"adresse\": \"1200 Frank ave\",\r\n    \"pseudo\": \"SparePart\",\r\n    \"motDePasse\": \"12345\"\r\n}"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Retourne le json de l'usager créé"
                    }
                },
                "servers": [{
                    "url": "http://localhost:8090"
                }]
            },
            "servers": [{
                "url": "http://localhost:8090"
            }]
        },
...
        "/usagers/Usager_id/commandes/": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "description": "Création d\\'une commande.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/body"
                            },
                            "examples": {
                                "0": {
                                    "value": "{\r\n    \"_id\": \"6099469e6f9a3823a054c85c\",\r\n    \"nom\": \"Proulx\",\r\n    \"prenom\": \"Laurent\",\r\n    \"adresse\": \"123 rue Albert\",\r\n    \"pseudo\": \"Plaurent\",\r\n    \"motDePasse\": \"12345\"\r\n}"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "La commande est créée."
                    }
                },
                "servers": [{
                    "url": "http://localhost:8090"
                }]
            },
            "servers": [{
                "url": "http://localhost:8090"
            }]
        },
        "/usagers/Usager_id/commandes/Commande_id/livreur": {
            "put": {
                "tags": [
                    "Commandes"
                ],
                "description": "Retourne le json de la commande.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/body_1"
                            },
                            "examples": {
                                "0": {
                                    "value": "{\r\n    \"_id\": \"609aa9232a275351cc20dd9d\",\r\n    \"nom\": \"McQueen\",\r\n    \"prenom\": \"Flash\",\r\n    \"Voiture\": \"Nascar\",\r\n    \"Quartier\": \"Dinoco\",\r\n    \"__v\": 0\r\n}"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retourne le json de la commande.",
                        "content": {
                            "text/html; charset=utf-8": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {}
                            }
                        }
                    }
                },
                "servers": [{
                    "url": "http://localhost:8090"
                }]
            },
            "servers": [{
                "url": "http://localhost:8090"
            }]
        }
    },
    "components": {
...
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://yourauthurl.com",
                        "scopes": {
                            "scope_name": "Enter your scopes here"
                        }
                    }
                }
            }
        }
    }
}